app/requirements.txt
fastapi
uvicorn
azure-identity
azure-data-tables
